PB-1200: Tidy up code.
